Implements general idea for device threads
Each device thread consists of a main object Device which extends Thread.
The Thread object has an lower level device object, e.g. Nidevice,
and a TriggerHandler. The actual device object is responsible for
creating data and put it in a Q. The TriggerHandler accesses the Q
from the low level device and performs triggering using trigger objects.
If the triger is curently running, the data is moved to a new Q,
which is open access from the top level Device object.