New compact NIDevice
Implementa a new class 'NIDevice' which has the required functionality to do simple
input measurements, contained in a single class.
This class is a thread object and can therefore only be excecuted once.
At the moment there is no possibility to pause data flow or reconfigure the settings
once the device has started. This might be added in the future.