Logging and minor bugfixes
At this state the data acquisition is working, but there are some flaws in the design.

If the trigger for a QHandler is active it will rapidly clear out the buffer. There will be no wait
for any new blocks to arrive, which causes the loop in QHandler to execute rapidly.
If we are unlucky this thread will never release the GIL, so there will never be any new blocks.
This would fill up the device hardware buffer since we never read, but the error seems to
get lost somewhere in the void...

On the other hand, if blocks are arriving very fast the TriggerHandler might process blocks
faster that the QHandler consumes them, e.g. if there is a buid up in the Q from the device,
which would cause blocks to be pushed out from the deque before the QHandler has a
chanse to process them.

Proposed solution: Move the trigger condition back to the trigger handler and move
the blocks from the deque to the Q immidietely for each block we process.
The QHandler can then actually wait for new blocks since it is processing a Q
and not a deque. There is also no possibility of dropping blocks since the trigger
condition is guaranteed to be checked for each incomming block, and all blocks
that are transferred between threads are done so in queues with infinite length.