Redo LevelDetector
The old level detector had to much Python code and used the GIL
too much, so the other threads (specifically the acquisition callback)
could not keep up for high sample rates.

The new detector uses an iir filter from scipy to do the detection, at the
cost of dropping separate attack and realese times. This is much faster
in execution time and is (I think) based on a C-function so it does not use
GIL all the time.

The old detector is kept for future use and reference with the name
LevelDetectorAttackRealese.