Move the circular buffer to the TriggerHandler
This should eliminate the risk that blocks are missed. Previously the
QHandler could miss blocks if the TriggerHandler put many of them in the buffer before
the QHandler had a chanse to run.
This also reduces the number of checks that the QHandler needs to do.
Previously it had to constantly check if there are blocks in the buffer, (using GIL)
but now it can simply wait for blocks to arrive in the Q (not using GIL).