Enable unscaled measurements
Allows for unscaled ints to be read from NI devices. This is controlled using the
`dtype` property of NIDevice.
Adds methods to get properties relating to unscaled reads in NIDevice:
`bit_depth`, `word_length`, `scaling_coeffs`, and `input_range`.
`scaling_coeffs` returns the calibrated polynomial coefficients to
convert unscaled ints to scaled floats.
NOTE: For national instruments devices this is the right way to do it!
The coefficients are created individually for each channel and is why NI
hardware is 'calibrated'.

The unscaled ints are converted to floats temporary for triggering porposes only,
see `handlers.py`, but only using the linear scaling part.

The calibration function also converts to floats if needed, so that the calibration
value will not depend on the data type.

One assumption in this implementation is that no device will need higher order terms,
and that the triggers will work without the constant shift.