Enable output in device framework
Adds a second Q and a second event to control it.
This renames some of the existing variables to more logical names.

The new implementation will work for both input and output or mixed
devices seamlessly. It is however up to concrete device implementations
to decide if the input and output are independent or synchronized.