Bugfixes for Windows
Since Windows uses a different method to start new processes, it is not possible to
initialize anything thread related before creating the new process.
This moves initializatoin of all intra-thread queues and events to the process target.

The process target is renamed so that it can be found in subclasses. Apparently the buitin
name wranglig is not reliable when multiprocessing.

Adds a `_hardware_setup` method that should be used to setup the hardware in the new process.