Implements framework for shared attributes
A framework for updating attributes cross processes is implemented using queues.
This can be used by declaring attributes as `core.InterProcessAttr` for properties that
should be shared between multiple processes. This relies on the hardware thread
calling `_update_attrs` to check for updates. This will check for requested changes
as well as check for requested values.

The framework is working at least to some extent, but it is still in
an experimental stage.