Implements serial device in new framework
Communication with serial devices i.e. signal generators is implemented in the new Device framework.
This allows controll over a serial generator using the triggering mechanisms implemented in the core module.
The commands are based on an Agilent 33120 A signal generator, but should be fairly standard.

This is an example of how to use the new `InterProcessAttr` to controll a device.