Removes `_hardware_setup` from Device framework
The method was strictly speaking not needed, since all the setup could
as well be done in _hardware_run. By not separating these, local variables
cound be used more easily.

The corresponding `_hardware_reset` method is kept for the moment. In
principle the resetting could be done in `_hardware_run` as well, but
this might require further testing. The difference is that the triggers
and Q threads will still be running when `_hardware_run` exits, but they
have stopped when `_hardware_reset` is called.