Removes multiprocessing
Using a separate process for each device is a mess.
- The sounddevice package has serious problems with multiple processes.
- Unable to access certain properties of NI devices.
- Large workarounds to implement simple serial devices.

The overall structure is fine, so the same model is kept.
All old code is still there, and can be reverted simple by swhitching
which line is commented. At some point the old implementation will be
removed, but it will require more testing to see that separate processes
are not needed.