Rename process related variables
Since we are no longer using multiple processes it is not reasonable
to name the variables/frunctions as such. Most variables now use variations
of `main`, but the former `__process` is now called `__main_thread`.
This is a compromise, but it seems more reasonable to call a process
'thread' than to call a thread 'process'.