Generator framework
Adds a framework for attaching generators to a device. The generators
will be called if the output is active to create a output frame.
It is the generators responsibility that the framesize is correct.
A generator can deliver as many channels as it wants, but the total
number of channels from all generators should match the number of output
channels in the device.