Implements noise generators
Noise based generators are implemented. The generator supports
many noise slopes, e.g. pink, white, and has two implementations.
The first implementation is a simple fft base generation, where
a white spectra is multiplied with appropriate coefficients.
The second method is based on an autoregressive model. This separates
the generation quality and the frame size nicely.