Moves channel handling to framework
A custom class for channels is created in the Device framework.
This gives the possibility to store more information, e.g. calibraitons,
in the same place as the channel index. The only downside is that
expressions which previously used the channel integers might now need
to explicitly convert to integers. See the change in `ni.py` for more
details.