Configures triggers to use calibrations and input scalings
Input scalings cal be used to compensate hardware mappings, e.g. using
int16 inputs instead of floats. This is a function which concreate
implementations can inplement to apply arbritrary scaling to inputs
before the input is passed to triggers.This is a once per device scaling.

Each trigger has a `use_calibration` property, which is checked upon
device setup. If true, the trigger will use calibrated values in the
implementation of the `test`.