Implements calibrations
This is not tested!
Due to lack of hardware on the platform where this was implmented,
the calibration method is not tested. The main uncertainty in this
implementation is the dynamic append and remove of the level detector
to the triggers. Since the triggering happens in a different thread,
race conditions might occur.