New implementation for serial signal generators
Since we are no longer using multiple processes,
we do not need a overly complicated setup to controll generators.
This new implementation if similar to the old `SerialInstrument`
but with access to more signal shapes, repeated sweeps,
and easier triggering in the triggering framework.

Be careful to not have a always on trigger call `on()`, since that will
put messages in the buffer faster than the generator can consume them.
This will generate a lot of hardware errors.