Sets a maaxsize for output Qs
The output Qs was filled very rapidly from unlimited generators, and
this also held the mutex locked and prevented the device from getting
data. A maxsize for the output Q ensures that the Q does not grow to
arbitrary length, and that the device has a chance to read from the Q.

If the device does not read from the Q, the generator thread will timeout
the call, check stop conditions, and try again with the same frame.