Makes `input_scaling` private
This method is used to scale inputs, e.g. from int32 to float64, but not
to account for calibration values. Its intended use is to provide scaled
values to triggers but still save the raw unscaled values to files.
Therefore it is not something that a user should interact with manually,
but developers might need to override it in implementations.