Lets `Channel` inherit from `int`
Instead of implementing each and every comparison, arithmetic, and index
simply inherit from int. This still allows for storage of attribures,
the json convertions, and nice printing, but otherwise behaves like an
integer.

This has the downside that two channels of different types will compare
as equal if the index are the same. Channels have mutable index but the
int part is immutable, so you could in principle change the index after
creation, but the int representation will have the original value.