Implements sample level trigger aignment
The implementation is tested only in simulations, where it is working as
intended. Further testing for actual hardware is needed.

Limitations: Currently the triggers signal the trigger alignment to the
device which is triggering, not the triggered device. This works good
when there is just a single device triggering itself, but will break
across multiple devices.

A possible method is to create alignment functions which sets the
alignment on the triggered device. This is relatively easy to do using
lambdas, but since it is such a common operation it could be nice to
simplify it somehow.