Changes trigger alignment for multiple devices
Instead of returning the trigger alignment to the triggering device,
each trigger will keep a list of devices to align and set the alignment
attribute on all of them. This scales the handover values to time instead
of samples, so it should work with different sample rates, as good as can
be expected. There is still desync problems if one device triggers
another device, since we are not actually syncronising the devices, only
soft-aligning.

A second parameter is added to the `add_trigger` function on devices,
which will specify the device which should be syncronised by this trigger.
This should normally be the same device which the trigger starts, but
there is no restrictions.