Notes on how Device should change
THe conceptual hierarchy of adding e.g. triggers to a device does not
really match how things work. It will be cleared if objects e.g. trigger
have a reference to the device and handes the device-side configuration
internally.

This would allow for lines like
```
dev = AudioDevice()
trig = RMSTrigger(level=0.6, device=dev)
...
```
which is cleaner than adding to the device.