Enables multiple modules in single NIDevice
Instead of representing a module with a device, we can represent a chassis with a device.
This allows for multiple modules to be used as a single multichannel device when
using a multi-slot chassis. If inputs and outputs are used simultaneously, the modules
will be syncronized using shared timebases. This is tested only using NI-9234 and NI-9260.