Implements padding and fading of measurement signals
The arbitrary signal generator is now capable of applying fade in and
fade out to the measurement signals, as well as zero-padding at both
ends. This is implemented as a wrapper around setting the signal,
so a user could create an empty `ArbitrarySignalGenerator` with
some fade values, and then simply set the signal to what they prefer.