Implements multiple deconvolution methods
All are based on calculating an inverse filter. Currently four methods
are implemented, each with it's own benefit.
The `filtering` method is similar to the previous implementation, with
mediocre performance.
The `regularization` method is conceptually similar to the filtering,
but regularizing the reference outside the swept range instead of
filtering. This seems to work better for the cases tested so far.
The `analytic` method does not really use a reference signal at all,
but relies on that the input is assumed to be en exponential sweep.
The inverse filter is calculated via an analytical expression for the
inverse spectrum.
The `time-reversal` method tries to do the same thing as the analytical
expression, but constructing the inverse filter as the time-reversed and
amplitude compensated excitation signal. Currently this does not work
well for the global amplitude, so no absolute measurements can be made.