Improve stability of NI outputs
Manually configuring the size of the output buffer allows us to increase
the stability of the entire system. Coupled with allowing the device
to reuse samples if we cannot manage to push them fast enough, will
prevent accidental breaks due to other python processing.
More inline comments in the code will hopefully improve the maintainability
of this somewhat difficult section.