Implements tracking of samplerates
The samplerate is tied to objects in the pipeline which have samplerates
defined from external sources. All other objects will access that
samplerate by keeping track of in which direction to look for it.
This should still allow for flexible arrangements in the future, by
implementing objects in the pipeline which do resampling or synchronization.