Implements Pipiline objects
We're running everything in pipilines anyhow, so we might as well have an
object for them. This sets the interface for pipelineing the nodes to
use the `|` symbol, implemented in the `__or__` methods.
This seems overall less error-prone in the implementation, since we
never work with more than two nodes at the time. Inserting nodes in
the middle of the pipeline is no longer possible, but that shold not be
a limitation.