New zarr recorder
The previous implementation was very strange with respect to gates,
storing each open segment as a new array. This implementation will instead
stack the "shots" along a new dimention in the data. Since zarr
will not store chunks that are fully empty, there won't be that much
wasted diskspace for unequal shots.